feat: Misc code tweaks and bug fixes (phase 4) (#25)

+++ api/src/modules/user/user.service.ts
@@ -4,6 +4,7 @@ import { Nullable } from 'src/common/types/nullable.type';
 import { ApiErrorCode } from 'src/exception/api-error-code.enum';
 import { ApiHttpException } from 'src/exception/api-http-exception.type';
 import { JwtPayload } from '../auth/core/strategies/jwt-payload.type';
+import { WorkspaceUserRole } from '../workspace/workspace-user-module/domain/workspace-user-role.enum';
 import { WorkspaceUserService } from '../workspace/workspace-user-module/workspace-user.service';
 import { User } from './domain/user.domain';
 import { RolePerWorkspace, UserResponse } from './dto/user-response.dto';
@@ -59,9 +60,9 @@ export class UserService {
   async createVirtualUser(
     data: Pick<User, 'firstName' | 'lastName' | 'status'>,
   ): Promise<User> {
-    const newVirtalUser = await this.userRepository.createVirtualUser(data);
+    const newVirtualUser = await this.userRepository.createVirtualUser(data);
 
-    if (!newVirtalUser) {
+    if (!newVirtualUser) {
       throw new ApiHttpException(
         {
           code: ApiErrorCode.SERVER_ERROR,
@@ -70,7 +71,7 @@ export class UserService {
       );
     }
 
-    return newVirtalUser;
+    return newVirtualUser;
   }
 
   async me(data: JwtPayload): Promise<UserResponse> {
@@ -97,17 +98,15 @@ export class UserService {
     return userDto;
   }
 
-  async findById(id: User['id']): Promise<Nullable<User>> {
+  findById(id: User['id']): Promise<Nullable<User>> {
     return this.userRepository.findById(id);
   }
 
-  async findByEmail(
-    email: NonNullable<User['email']>,
-  ): Promise<Nullable<User>> {
+  findByEmail(email: NonNullable<User['email']>): Promise<Nullable<User>> {
     return this.userRepository.findByEmail(email);
   }
 
-  async findBySocialIdAndProvider({
+  findBySocialIdAndProvider({
     socialId,
     provider,
   }: {
@@ -158,18 +157,57 @@ export class UserService {
     });
 
     if (!updatedUser) {
+      // Somebody deleted themselves in the meantime
       throw new ApiHttpException(
         {
-          code: ApiErrorCode.SERVER_ERROR,
+          code: ApiErrorCode.INVALID_PAYLOAD,
         },
-        HttpStatus.INTERNAL_SERVER_ERROR,
+        HttpStatus.NOT_FOUND,
       );
     }
 
     return updatedUser;
   }
 
+  /**
+   * This should cascadely delete workspace_user,
+   * goal, task_assignment and session entities
+   * and set NULL to specific entities' properties.
+   */
   async delete(userId: User['id']): Promise<void> {
-    await this.userRepository.delete(userId);
+    // We firstly need to check if user is the last Manager
+    // in any workspace. Because if he was, deleting
+    // it immediately would leave that/those workspace/s
+    // in a state where it can't be admistrated anymore.
+    const workspaceUsers =
+      await this.workspaceUserService.findAllByUserId(userId);
+    const workspaceUsersLastManager = workspaceUsers.filter(
+      (wu) => wu.workspaceRole === WorkspaceUserRole.MANAGER,
+    );
+
+    // If there are some workspaces user is part of and is Manager in them
+    // we block the deletion, and send those workspaces in the response
+    // and prompt the user to delete them before deleting the account.
+    if (workspaceUsersLastManager.length > 0) {
+      throw new ApiHttpException(
+        {
+          code: ApiErrorCode.SOLE_MANAGER_CONFLICT,
+        },
+        HttpStatus.CONFLICT,
+      );
+    }
+
+    const result = await this.userRepository.delete(userId);
+
+    if (!result) {
+      // Should never happen because we read userId from JWT
+      // and if JWT is invalid, it will fail on the JWT guard
+      throw new ApiHttpException(
+        {
+          code: ApiErrorCode.INVALID_PAYLOAD,
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
   }
 }
